refactor: деньги и даты хранятся точными типами
Балансы, средние цены и комиссии были double в моделях и
DOUBLE PRECISION в схеме — копейки дрейфовали при каждом накоплении,
а сравнения сумм зависели от порядка операций. Даты (baseline-даты,
last_replenish_date) хранились строками.

- модели: cash, borrowed, total_fees, avg_price, price, fee,
  limit_price, trigger_price, target_price, amount_rub и валютные
  остатки — BigDecimal;
- схема: NUMERIC(19,8) для денег, DATE для дат, с идемпотентной
  миграцией ALTER ... USING в schema.sql;
- репозитории читают getBigDecimal/getObject(LocalDate) вместо
  getDouble и ручного LocalDate.parse;
- BaseRepository.nz() приводит NULL к нулю: getDouble() возвращал 0.0
  сам, getBigDecimal() отдаёт null, и старая запись с пустым полем
  иначе роняла бы расчёты.

Проверено на копии боевой БД, поднятой из свежего pg_dump: типы
сменились, балансы, позиции и сделки на месте, повторный прогон
schema.sql проходит без ошибок.

ВАЖНО: после миграции старый jar с БД не работает — откат деплоя
потребует восстановления дампа
(/mnt/storage/backups/postgres/stonks_pre_migration_20260902_1151.sql).

+++ src/main/java/services/sandbox/SandboxOrderProcessor.java
@@ -78,7 +78,7 @@ public class SandboxOrderProcessor {
 			BigDecimal price = prices.getOrDefault(share.getUid(), BigDecimal.ZERO);
 			if (price.compareTo(BigDecimal.ZERO) <= 0) continue;
 
-			BigDecimal triggerPrice = BigDecimal.valueOf(so.getTriggerPrice());
+			BigDecimal triggerPrice = so.getTriggerPrice();
 			boolean triggered = so.getType() == StopOrderType.SL
 					? price.compareTo(triggerPrice) <= 0
 					: price.compareTo(triggerPrice) >= 0;
@@ -128,7 +128,7 @@ public class SandboxOrderProcessor {
 			BigDecimal price = prices.getOrDefault(share.getUid(), BigDecimal.ZERO);
 			if (price.compareTo(BigDecimal.ZERO) <= 0) continue;
 
-			BigDecimal loLimitPrice = BigDecimal.valueOf(lo.getLimitPrice());
+			BigDecimal loLimitPrice = lo.getLimitPrice();
 			boolean isBuy = lo.getSide() == TradeSide.BUY;
 			boolean triggered = isBuy
 					? price.compareTo(loLimitPrice) <= 0
@@ -172,7 +172,7 @@ public class SandboxOrderProcessor {
 			BigDecimal price = prices.getOrDefault(share.getUid(), BigDecimal.ZERO);
 			if (price.compareTo(BigDecimal.ZERO) <= 0) continue;
 
-			BigDecimal alertTarget = BigDecimal.valueOf(alert.getTargetPrice());
+			BigDecimal alertTarget = alert.getTargetPrice();
 			boolean triggered = alert.isAbove()
 					? price.compareTo(alertTarget) >= 0
 					: price.compareTo(alertTarget) <= 0;
